+++ src/api/BdmsContext.cs
@@ -25,6 +25,7 @@ public class BdmsContext : DbContext
     public DbSet<UserWorkgroupRole> UserWorkgroupRoles { get; set; }
     public DbSet<Workflow> Workflows { get; set; }
     public DbSet<Workgroup> Workgroups { get; set; }
+    public DbSet<BoreholeFile> BoreholeFiles { get; set; }
 
     public BdmsContext(DbContextOptions options)
         : base(options)
@@ -37,5 +38,19 @@ public class BdmsContext : DbContext
     {
         modelBuilder.HasDefaultSchema("bdms");
         modelBuilder.Entity<UserWorkgroupRole>().HasKey(k => new { k.UserId, k.WorkgroupId, k.Role });
+
+        modelBuilder.Entity<Borehole>()
+                .HasMany(b => b.Files)
+                .WithMany(f => f.Boreholes)
+                .UsingEntity<BoreholeFile>(
+                    j => j
+                        .HasOne(bf => bf.File)
+                        .WithMany(f => f.BoreholeFiles)
+                        .HasForeignKey(bf => bf.FileId),
+                    j => j
+                        .HasOne(bf => bf.Borehole)
+                        .WithMany(b => b.BoreholeFiles)
+                        .HasForeignKey(bf => bf.BoreholeId),
+                    j => j.HasKey(bf => new { bf.BoreholeId, bf.FileId }));
     }
 }